+++ packages/pob-dependencies/index.d.ts
@@ -5,11 +5,11 @@ interface Dependencies {
   '@babel/preset-react': '7.0.0';
   '@commitlint/cli': '7.5.0';
   '@commitlint/config-conventional': '7.5.0';
-  '@types/jest': '23.3.13';
+  '@types/jest': '24.0.0';
   '@typescript-eslint/eslint-plugin': '1.2.0';
-  '@typescript-eslint/parser': '1.2.0';
+  '@typescript-eslint/parser': '1.3.0';
   'babel-eslint': '10.0.1';
-  'babel-jest': '24.0.0';
+  'babel-jest': '24.1.0';
   'babel-preset-latest-node': '3.0.1';
   'babel-preset-modern-browsers': '13.0.1';
   'eslint': '5.13.0';
@@ -17,20 +17,20 @@ interface Dependencies {
   'eslint-config-pob': '26.0.8';
   'eslint-plugin-babel': '5.3.0';
   'eslint-plugin-import': '2.16.0';
-  'eslint-plugin-jsx-a11y': '6.2.0';
+  'eslint-plugin-jsx-a11y': '6.2.1';
   'eslint-plugin-node': '8.0.1';
   'eslint-plugin-react': '7.12.4';
   'husky': '1.3.1';
-  'jest': '24.0.0';
+  'jest': '24.1.0';
   'jest-junit-reporter': '1.1.0';
-  'lerna': '3.10.8';
+  'lerna': '3.11.0';
   'lint-staged': '8.1.3';
-  'pob-babel': '^23.6.0';
+  'pob-babel': '^23.6.1';
   'pob-release': '^4.7.0';
-  'prettier': '1.16.3';
+  'prettier': '1.16.4';
   'typedoc': '0.14.2';
-  'typescript': '3.3.1';
+  'typescript': '3.3.3';
   'yarnhook': '0.3.0';
-}
+} 
 
 export = dependencies;