+++ packages/pob-babel/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [23.6.2](https://github.com/christophehurpeau/pob/compare/pob-babel@23.6.1...pob-babel@23.6.2) (2019-02-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency prettier to v1.16.4 ([#88](https://github.com/christophehurpeau/pob/issues/88)) ([d33897a](https://github.com/christophehurpeau/pob/commit/d33897a))
+
+
+
+
+
 ## [23.6.1](https://github.com/christophehurpeau/pob/compare/pob-babel@23.6.0...pob-babel@23.6.1) (2019-02-02)